requirements.txt
reflex==0.4.0
opencv-python==4.9.0.80
easyocr==1.7.1
matplotlib==3.8.2
SpeechRecognition==3.10.1
pillow==10.2.0
pygame==2.5.2
PyAudio==0.2.14
